:recycle: Factor out file update tests
